requirements.txt
Pillow>=9.4.0
requests>=2.21.0
icmplib>=3.0.3